+++ skysail.api/src/io/skysail/api/um/AuthenticationService.java
@@ -7,18 +7,44 @@ import org.restlet.Context;
 import org.restlet.Request;
 import org.restlet.security.Authenticator;
 
+/**
+ * Defines the contract for authentication handling in skysail applications.
+ */
 @ProviderType
 public interface AuthenticationService {
 
-	Authenticator getApplicationAuthenticator(Context context);
+    /**
+     * @return a restlet authenticator the application is authenticated against.
+     */
+    Authenticator getApplicationAuthenticator(Context context);
 
-	Authenticator getResourceAuthenticator(Context context);
+    /**
+     * @return a restlet authenticator the applications' resources are
+     *         authenticated against.
+     */
+    Authenticator getResourceAuthenticator(Context context);
 
-	boolean isAuthenticated(Request request);
+    /**
+     * @return whether or not the current request is authenticated.
+     */
+    boolean isAuthenticated(Request request);
 
-	Principal getPrincipal(Request request);
+    /**
+     * @return the associated principal, i.e. the currently authenticated user.
+     */
+    Principal getPrincipal(Request request);
 
-	String getLoginPath();
+    /**
+     * skysail can be set up to run with different authentication services, which
+     * typically provide a login path to which the generic "/_login" path will
+     * be redirected.
+     */
+    String getLoginPath();
 
-	String getLogoutPath();
+    /**
+     * skysail can be set up to run with different authentication services, which
+     * typically provide a logout path to which the generic "/_logout" path will
+     * be redirected.
+     */
+    String getLogoutPath();
 }