+++ skysail.api/src/io/skysail/api/um/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Root of the user management (UM) API.
+ *
+ * This package defines the interfaces of skysail related to usermanagement,
+ * authentication and authorization.
+ *
+ */
+@Version("2.0.0")
+package io.skysail.api.um;
+
+import org.osgi.annotation.versioning.Version;