+++ skysail.api/etc/changelog.md
@@ -1,10 +1,11 @@
 new in
 
-2.2.0
+3.0.0
 ------
 
  * added getLoginPath() and getLogoutPath() methods to Authentication Service
  * removed biz.aQute.bnd.annotations
+ * user management cleanup and documentation
 
 2.1.0
 ------